refactor(storage-proofs): move non-unit tests into 'tests' folders
