+++ storage-proofs/core/src/merkle/proof.rs
@@ -101,10 +101,6 @@ pub trait MerkleProofTrait:
     fn expected_len(&self, leaves: usize) -> usize {
         compound_path_length::<Self::Arity, Self::SubTreeArity, Self::TopTreeArity>(leaves)
     }
-
-    /// Test only method to break a valid proof.
-    #[cfg(test)]
-    fn break_me(&mut self, leaf: <Self::Hasher as Hasher>::Domain);
 }
 
 pub fn base_path_length<A: Unsigned, B: Unsigned, C: Unsigned>(leaves: usize) -> usize {
@@ -297,22 +293,6 @@ impl<
     }
     fn path_index(&self) -> usize {
         forward_method!(self.data, path_index)
-    }
-
-    /// Test only method to break a valid proof.
-    #[cfg(test)]
-    fn break_me(&mut self, leaf: H::Domain) {
-        match self.data {
-            ProofData::Single(ref mut proof) => {
-                proof.leaf = leaf;
-            }
-            ProofData::Sub(ref mut proof) => {
-                proof.leaf = leaf;
-            }
-            ProofData::Top(ref mut proof) => {
-                proof.leaf = leaf;
-            }
-        }
     }
 }
