+++ storage-proofs/post/src/election/compound.rs
@@ -2,9 +2,7 @@ use std::marker::PhantomData;
 
 use bellperson::bls::{Bls12, Fr};
 use bellperson::Circuit;
-use generic_array::typenum;
-use typenum::marker_traits::Unsigned;
-
+use generic_array::typenum::Unsigned;
 use storage_proofs_core::{
     compound_proof::{CircuitComponent, CompoundProof},
     drgraph,
@@ -156,159 +154,5 @@ where
             sector_id: None,
             _t: PhantomData,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use std::collections::BTreeMap;
-
-    use rand::SeedableRng;
-    use rand_xorshift::XorShiftRng;
-
-    use bellperson::util_cs::{metric_cs::MetricCS, test_cs::TestConstraintSystem};
-    use filecoin_hashers::{poseidon::PoseidonHasher, Domain, HashFunction, Hasher};
-    use storage_proofs_core::{
-        compound_proof,
-        merkle::{generate_tree, get_base_tree_count, LCTree, MerkleTreeTrait},
-        proof::NoRequirements,
-        sector::SectorId,
-    };
-    use typenum::{U0, U8};
-
-    use crate::election;
-
-    #[ignore] // Slow test – run only when compiled for release.
-    #[test]
-    fn election_post_test_compound_poseidon() {
-        election_post_test_compound::<LCTree<PoseidonHasher, U8, U0, U0>>();
-    }
-
-    fn election_post_test_compound<Tree: 'static + MerkleTreeTrait>() {
-        let rng = &mut XorShiftRng::from_seed(crate::TEST_SEED);
-
-        let leaves = 64 * get_base_tree_count::<Tree>();
-        let sector_size = (leaves * NODE_SIZE) as u64;
-        let randomness = <Tree::Hasher as Hasher>::Domain::random(rng);
-        let prover_id = <Tree::Hasher as Hasher>::Domain::random(rng);
-
-        let setup_params = compound_proof::SetupParams {
-            vanilla_params: election::SetupParams {
-                sector_size,
-                challenge_count: 20,
-                challenged_nodes: 1,
-            },
-            partitions: None,
-            priority: true,
-        };
-
-        let mut sectors: Vec<SectorId> = Vec::new();
-        let mut trees = BTreeMap::new();
-
-        let temp_dir = tempfile::tempdir().unwrap();
-        let temp_path = temp_dir.path();
-
-        for i in 0..5 {
-            sectors.push(i.into());
-            let (_data, tree) =
-                generate_tree::<Tree, _>(rng, leaves, Some(temp_path.to_path_buf()));
-            trees.insert(i.into(), tree);
-        }
-
-        let pub_params = ElectionPoStCompound::<Tree>::setup(&setup_params).expect("setup failed");
-
-        let candidates = election::generate_candidates::<Tree>(
-            &pub_params.vanilla_params,
-            &sectors,
-            &trees,
-            prover_id,
-            randomness,
-        )
-        .unwrap();
-
-        let candidate = &candidates[0];
-        let tree = trees.remove(&candidate.sector_id).unwrap();
-        let comm_r_last = tree.root();
-        let comm_c = <Tree::Hasher as Hasher>::Domain::random(rng);
-        let comm_r = <Tree::Hasher as Hasher>::Function::hash2(&comm_c, &comm_r_last);
-
-        let pub_inputs = election::PublicInputs {
-            randomness,
-            sector_id: candidate.sector_id,
-            prover_id,
-            comm_r,
-            partial_ticket: candidate.partial_ticket,
-            sector_challenge_index: 0,
-        };
-
-        let priv_inputs = election::PrivateInputs::<Tree> {
-            tree,
-            comm_c,
-            comm_r_last,
-        };
-
-        {
-            let (circuit, inputs) =
-                ElectionPoStCompound::circuit_for_test(&pub_params, &pub_inputs, &priv_inputs)
-                    .unwrap();
-
-            let mut cs = TestConstraintSystem::new();
-
-            circuit.synthesize(&mut cs).expect("failed to synthesize");
-
-            if !cs.is_satisfied() {
-                panic!(
-                    "failed to satisfy: {:?}",
-                    cs.which_is_unsatisfied().unwrap()
-                );
-            }
-            assert!(
-                cs.verify(&inputs),
-                "verification failed with TestContraintSystem and generated inputs"
-            );
-        }
-
-        // Use this to debug differences between blank and regular circuit generation.
-        {
-            let (circuit1, _inputs) =
-                ElectionPoStCompound::circuit_for_test(&pub_params, &pub_inputs, &priv_inputs)
-                    .unwrap();
-            let blank_circuit =
-                ElectionPoStCompound::<Tree>::blank_circuit(&pub_params.vanilla_params);
-
-            let mut cs_blank = MetricCS::new();
-            blank_circuit
-                .synthesize(&mut cs_blank)
-                .expect("failed to synthesize");
-
-            let a = cs_blank.pretty_print_list();
-
-            let mut cs1 = TestConstraintSystem::new();
-            circuit1.synthesize(&mut cs1).expect("failed to synthesize");
-            let b = cs1.pretty_print_list();
-
-            for (i, (a, b)) in a.chunks(100).zip(b.chunks(100)).enumerate() {
-                assert_eq!(a, b, "failed at chunk {}", i);
-            }
-        }
-        let blank_groth_params =
-            ElectionPoStCompound::<Tree>::groth_params(Some(rng), &pub_params.vanilla_params)
-                .expect("failed to generate groth params");
-
-        let proof = ElectionPoStCompound::prove(
-            &pub_params,
-            &pub_inputs,
-            &priv_inputs,
-            &blank_groth_params,
-        )
-        .expect("failed while proving");
-
-        let verified =
-            ElectionPoStCompound::verify(&pub_params, &pub_inputs, &proof, &NoRequirements)
-                .expect("failed while verifying");
-
-        assert!(verified);
     }
 }
