+++ storage-proofs/post/src/lib.rs
@@ -4,8 +4,3 @@
 pub mod election;
 pub mod fallback;
 pub mod rational;
-
-#[cfg(test)]
-pub(crate) const TEST_SEED: [u8; 16] = [
-    0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc, 0xe5,
-];
